app/requirements.txt
asyncpg==0.27.0
SQLAlchemy==2.0.9
aiohttp==3.8.4
cachetools==5.3.0
python-dotenv==1.0.0
certifi